+++ Editor/Attributes/Utilities/PropertyUtility.cs
@@ -19,7 +19,7 @@ namespace TG.Attributes.Editor {
         
         public static System.Type GetPropertyType(SerializedProperty property) {
             var type = property.serializedObject.targetObject.GetType();
-            var field = type.GetField(property.propertyPath, BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance);
+            var field = type.GetField(property.propertyPath, BindingFlags.Public | BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.Static);
             return field.FieldType;
         }
     }