+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 - Fixed a bug where enum values couldn't be assigned for conditional attributes
 - Fixed a bug where if a scriptable object was empty it would throw an error in the drawer
-
+- Fixed issues with private fields not being displayed in the drawer
 ## [1.0.0] - 30/07/2024
 ### Added
 - Initial release